Fix for #5895 - Changed deprecated numpy.float to float (#5897)
* Removed numpy.float (deprecated) and coverted to float

* Removed numpy.float (deprecated) and coverted to float

+++ rdkit/ML/Cluster/Resemblance.py
@@ -45,7 +45,7 @@ def EuclideanDistance(inData):
 
   """
   nObjs = len(inData)
-  res = numpy.zeros((nObjs * (nObjs - 1) / 2), numpy.float)
+  res = numpy.zeros((nObjs * (nObjs - 1) / 2), float)
   nSoFar = 0
   for col in range(1, nObjs):
     for row in range(col):
